+++ packages/yourloops/test/integration/pages/patient-view/patient-profile/patient-profile-view-mobile.hcp.spec.tsx
@@ -25,52 +25,56 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { act } from 'react'
 import { mockAuth0Hook } from '../../../mock/auth0.hook.mock'
 import { mockNotificationAPI } from '../../../mock/notification.api.mock'
-import { patient1Id } from '../../../data/patient.api.data'
 import { mockDirectShareApi } from '../../../mock/direct-share.api.mock'
-import { renderPage } from '../../../utils/render'
-import { mockUserApi } from '../../../mock/user.api.mock'
-import { mockPatientApiForCaregivers } from '../../../mock/patient.api.mock'
-import { UserRole } from '../../../../../lib/auth/models/enums/user-role.enum'
-import { mockDataAPI, oneDayDashboardData } from '../../../mock/data.api.mock'
 import {
-  type AppMainLayoutParams,
-  testAppMainLayoutForCaregiverMobile
-} from '../../../use-cases/app-main-layout-visualisation'
+  mockTeamAPI,
+  myThirdTeamId,
+} from '../../../mock/team.api.mock'
+import { mockUserApi } from '../../../mock/user.api.mock'
+import { mockPatientApiForHcp } from '../../../mock/patient.api.mock'
+import { mockDataAPI } from '../../../mock/data.api.mock'
+import { renderPage } from '../../../utils/render'
+import { patient1Id } from '../../../data/patient.api.data'
 import { AppUserRoute } from '../../../../../models/enums/routes.enum'
-import { PRIVATE_TEAM_ID } from '../../../../../lib/team/team.util'
+import { act } from '@testing-library/react'
 import { mockDblCommunicationApi } from '../../../mock/dbl-communication.api'
+import { mockLeadCliniciansApi } from '../../../mock/clinicians.api.mock'
+import { mockErrorApi } from '../../../mock/error.api.mock'
+import { mockAnalyticsApi } from '../../../mock/analytics.api.mock'
+import { checkHCPAndCaregiverHeaderPatientViewMobile } from '../../../assert/header-mobile.assert'
 import { mockMobileScreen } from '../../../mock/mobile-screen.mock'
 
-describe('Dashboard view for caregiver', () => {
-  const patientDashboardRoute = `/teams/${PRIVATE_TEAM_ID}/patients/${patient1Id}${AppUserRoute.Dashboard}`
-  const firstName = 'Caregiver firstName'
-  const lastName = 'Caregiver lastName'
-
+describe('Patient profile view for HCP', () => {
   beforeEach(() => {
-    mockAuth0Hook(UserRole.Caregiver)
-    mockDblCommunicationApi()
+    mockAuth0Hook()
     mockNotificationAPI()
     mockDirectShareApi()
+    mockTeamAPI()
     mockUserApi().mockUserDataFetch({ firstName, lastName })
-    mockPatientApiForCaregivers()
+    mockPatientApiForHcp()
+    mockDataAPI()
+    mockDblCommunicationApi()
+    mockLeadCliniciansApi()
+    mockErrorApi()
+    mockAnalyticsApi()
     mockMobileScreen()
   })
 
-  it('In mobile version, should render correct components', async () => {
-    mockDataAPI(oneDayDashboardData)
-    const appMainLayoutParams: AppMainLayoutParams = {
-      footerHasLanguageSelector: false,
-      loggedInUserFullName: `${lastName} ${firstName}`
-    }
+  const firstName = 'HCP firstName'
+  const lastName = 'HCP lastName'
 
-    await act(async () => {
-      renderPage(patientDashboardRoute)
+  const patientProfileRoute = `/teams/${myThirdTeamId}/patients/${patient1Id}${AppUserRoute.PatientProfile}`
+
+  describe('Alerts section', () => {
+    it('should render correct layout', async () => {
+      await act(async () => {
+        renderPage(patientProfileRoute)
+      })
+
+      checkHCPAndCaregiverHeaderPatientViewMobile(`${lastName} ${firstName}`)
     })
 
-    await testAppMainLayoutForCaregiverMobile(appMainLayoutParams)
   })
-
 })